Movido el AuthController a un package de security.
