+++ src/main/java/Javastral/com/gestorMateriasWeb/security/controller/AuthController.java
@@ -1,0 +1,35 @@
+package Javastral.com.gestorMateriasWeb.security.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import Javastral.com.gestorMateriasWeb.security.payload.LoginRequest;
+import Javastral.com.gestorMateriasWeb.security.payload.SignupRequest;
+import Javastral.com.gestorMateriasWeb.security.service.AuthService;
+import jakarta.validation.Valid;
+
+@CrossOrigin(origins = "*", maxAge = 3600)
+@RestController
+@RequestMapping("/api/auth")
+public class AuthController {
+
+	@Autowired
+	private AuthService authService;
+	
+	@PostMapping("/signin")
+	public ResponseEntity<?> authenticateUser(@Valid @RequestBody LoginRequest loginRequest) {
+		return ResponseEntity.ok(
+				this.authService.signIn(loginRequest)
+				);
+	}
+
+	@PostMapping("/signup")
+	public ResponseEntity<?> registerUser(@Valid @RequestBody SignupRequest signUpRequest) {
+		return this.authService.signUp(signUpRequest);
+	}
+}
